wip(cwrp): deprecate crwp.
Server-side tests are working, but browser-side tests are really broken.

Main issue is that setting props on a router should fire an onNavigate() action,
but it doesn't since there is no longer state and no code runs on a prop change.

We may have to move this callback to render() somehow, or rethink how we handle
prop changes completely.

We also may have to remove the user-supplied `setRoutingState`.